+++ ADP/Fusion.hs
@@ -29,6 +29,7 @@ module ADP.Fusion
   , module ADP.Fusion.None
 --  , module ADP.Fusion.Region
   , module ADP.Fusion.Table
+  , module ADP.Fusion.Table.Fill
   , module ADP.Fusion.TH
   ) where
 
@@ -43,6 +44,7 @@ import           ADP.Fusion.Multi.Classes
 import           ADP.Fusion.None
 --import           ADP.Fusion.Region
 import           ADP.Fusion.Table
+import           ADP.Fusion.Table.Fill
 import           ADP.Fusion.TH
 
 import           Data.Array.Repa.Index