+++ ADP/Fusion/Examples/NeedlemanWunsch.hs
@@ -2,6 +2,16 @@
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+-- | The Needleman-Wunsch global alignment algorithm. This algorithm is
+-- extremely simple but provides a good showcase for what ADPfusion offers.
+--
+-- We automatically generate the algebra product using
+-- 'makeAlgebraProductH'.
+--
+-- Table filling and freezing is done automatically using 'runFreezeMTbls'
+-- as can be seen in the 'forwardPhase' function.
 
 module Main where
 
@@ -16,18 +26,9 @@ import qualified Data.Vector.Fusion.Stream.Monadic as S
 import qualified Data.Vector.Unboxed as VU
 import           System.IO.Unsafe (unsafePerformIO)
 
-import Data.Array.Repa.Index.Points
-import qualified Data.PrimitiveArray as PA
-import qualified Data.PrimitiveArray.Zero as PA
-import qualified Data.PrimitiveArray.FillTables as PA
+import           Data.PrimitiveArray as PA hiding (map)
 
-import ADP.Fusion.Chr
-import ADP.Fusion.Empty hiding (empty)
-import ADP.Fusion hiding (empty)
-import ADP.Fusion.Multi
-import ADP.Fusion.Table
-import ADP.Fusion.TH
-import Data.Array.Repa.Index.Subword
+import           ADP.Fusion
 
 
 
@@ -91,17 +92,19 @@ needlemanWunsch i1 i2 = (ws PA.! (Z:.pointL 0 n1:.pointL 0 n2), bt) where
   bt = take 1 $ backtrack i1 i2 (Z:.ws)
 {-# NOINLINE needlemanWunsch #-}
 
-forwardPhase :: VU.Vector Char -> VU.Vector Char -> IO (Z:.(PA.Unboxed (Z:.PointL:.PointL) Int))
+type Arr  = PA.Unboxed (Z:.PointL:.PointL) Int
+type Arrs = Z:.Arr
+
+forwardPhase :: VU.Vector Char -> VU.Vector Char -> IO Arrs
 forwardPhase i1 i2 = do
   let n1 = VU.length i1
   let n2 = VU.length i2
   !t' <- PA.newWithM (Z:.pointL 0 0:.pointL 0 0) (Z:.pointL 0 n1:.pointL 0 n2) (-999999)
-  let t= mTblD (Z:.EmptyOk:.EmptyOk) t'
-  PA.runFillTables . expose $ grammar sScore t i1 i2
-  PA.freeze t' >>= return . (Z:.)
+  let t = mTblD (Z:.EmptyOk:.EmptyOk) t'
+  runFreezeMTbls $ grammar sScore t i1 i2
 {-# INLINE forwardPhase #-}
 
-backtrack :: VU.Vector Char -> VU.Vector Char -> (Z:.(PA.Unboxed (Z:.PointL:.PointL) Int)) -> [[String]]
+backtrack :: VU.Vector Char -> VU.Vector Char -> Arrs -> [[String]]
 backtrack i1 i2 (Z:.t') = map (map reverse) . unId . S.toList . unId . g $ Z:.pointL 0 n1 :. pointL 0 n2 where
   n1 = VU.length i1
   n2 = VU.length i2